pyproject.toml
[tool.poetry]
name = "cookiecutter_project_name"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[[tool.poetry.source]]
name = "default"
url = "https://pypi.org/simple"
default = true

[[tool.poetry.source]]
name = "srvhub"
url = "http://pypi.srvhub.tools/simple"
secondary = true

[tool.poetry.dependencies]
python = "^3.9"
django = "^3.2.7"
requests = "^2.26.0"
psycopg2 = "^2.9.1"
djangorestframework = "^3.12.4"
djangorestframework-simplejwt = "^4.8.0"
django-environ = "^0.7.0"
appversion = { version = "^0.1.3", source = "srvhub" }
sentry-sdk = "^1.3.1"
django-redis = "^5.0.0"
pymemcache = "^3.5.0"
dj-model-utils = { version = "^0.1.0", source = "srvhub" }
djnewsletter = { version = "^2.3.0", source = "srvhub" }
celery = "^5.1.2"
drf-yasg = { version = "^1.20.0", extras = ["validation"] }

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
